+++ packages/dx-grid-core/src/plugins/grouping-state/helpers.test.js
@@ -1,5 +1,5 @@
 import {
-    getFirstChangedGropingIndex,
+  getFirstChangedGropingIndex,
 } from './helpers';
 
 describe('GroupingPlugin helpers', () => {