+++ packages/dx-grid-core/src/plugins/selection-state/computeds.test.js
@@ -1,6 +1,6 @@
 import {
-    getAvailableToSelect,
-    getAvailableSelection,
+  getAvailableToSelect,
+  getAvailableSelection,
 } from './computeds';
 
 describe('PagingState computeds', () => {