+++ packages/dx-grid-core/src/plugins/table-group-row/helpers.js
@@ -6,3 +6,4 @@ export const isGroupTableCell = (tableRow, tableColumn) =>
 export const isGroupIndentTableCell = (tableRow, tableColumn) =>
   tableRow.type === TABLE_GROUP_TYPE && tableColumn.type === TABLE_GROUP_TYPE
   && tableRow.row.groupedBy !== tableColumn.column.name;
+export const isGroupTableRow = tableRow => tableRow.type === TABLE_GROUP_TYPE;