+++ packages/dx-react-grid-material-ui/src/plugins/table-edit-row.jsx
@@ -3,11 +3,14 @@ import PropTypes from 'prop-types';
 import { combineTemplates } from '@devexpress/dx-react-core';
 import { TableEditRow as TableEditRowBase } from '@devexpress/dx-react-grid';
 import { EditCell } from '../templates/table-edit-cell';
+import { TableRow } from '../templates/table-row';
 
 const defaultEditCellTemplate = props => <EditCell {...props} />;
+const defaultEditRowTemplate = props => <TableRow {...props} />;
 
 export const TableEditRow = ({
   editCellTemplate,
+  editRowTemplate,
   ...restProps
 }) => (
   <TableEditRowBase
@@ -15,12 +18,18 @@ export const TableEditRow = ({
       editCellTemplate,
       defaultEditCellTemplate,
     )}
+    editRowTemplate={combineTemplates(
+      editRowTemplate,
+      defaultEditRowTemplate,
+    )}
     {...restProps}
   />
 );
 TableEditRow.propTypes = {
   editCellTemplate: PropTypes.func,
+  editRowTemplate: PropTypes.func,
 };
 TableEditRow.defaultProps = {
   editCellTemplate: undefined,
+  editRowTemplate: undefined,
 };