+++ packages/dx-react-grid-material-ui/src/plugins/table-row-detail.jsx
@@ -1,16 +1,46 @@
 import React from 'react';
+import PropTypes from 'prop-types';
+import { combineTemplates } from '@devexpress/dx-react-core';
 import { TableRowDetail as TableRowDetailBase } from '@devexpress/dx-react-grid';
 import { TableDetailToggleCell } from '../templates/table-detail-toggle-cell';
 import { TableDetailCell } from '../templates/table-detail-cell';
+import { TableRow } from '../templates/table-row';
 
-const detailToggleCellTemplate = props => <TableDetailToggleCell {...props} />;
-const detailCellTemplate = props => <TableDetailCell {...props} />;
+const defaultDetailToggleCellTemplate = props => <TableDetailToggleCell {...props} />;
+const defaultDetailCellTemplate = props => <TableDetailCell {...props} />;
+const defaultDetailRowTemplate = props => <TableRow {...props} />;
 
-export const TableRowDetail = props => (
+export const TableRowDetail = ({
+  detailCellTemplate,
+  detailRowTemplate,
+  detailToggleCellTemplate,
+  ...restProps
+}) => (
   <TableRowDetailBase
-    detailToggleCellTemplate={detailToggleCellTemplate}
-    detailCellTemplate={detailCellTemplate}
+    detailToggleCellTemplate={combineTemplates(
+      detailToggleCellTemplate,
+      defaultDetailToggleCellTemplate,
+    )}
+    detailCellTemplate={combineTemplates(
+      detailCellTemplate,
+      defaultDetailCellTemplate,
+    )}
+    detailRowTemplate={combineTemplates(
+      detailRowTemplate,
+      defaultDetailRowTemplate,
+    )}
     detailToggleCellWidth={38}
-    {...props}
+    {...restProps}
   />
 );
+
+TableRowDetail.propTypes = {
+  detailRowTemplate: PropTypes.func,
+  detailCellTemplate: PropTypes.func,
+  detailToggleCellTemplate: PropTypes.func,
+};
+TableRowDetail.defaultProps = {
+  detailRowTemplate: undefined,
+  detailCellTemplate: undefined,
+  detailToggleCellTemplate: undefined,
+};