+++ packages/dx-react-grid/src/plugins/table-group-row.jsx
@@ -13,6 +13,7 @@ import {
   tableRowsWithGrouping,
   isGroupTableCell,
   isGroupIndentTableCell,
+  isGroupTableRow,
 } from '@devexpress/dx-grid-core';
 
 const getGroupIndentTableCellTemplateArgs = ({ params }) => ({
@@ -39,6 +40,11 @@ const getValueFormatterArgs = params => ({
   value: params.row.value,
 });
 
+const getGroupTableRowTemplateArgs = params => ({
+  ...params,
+  row: params.tableRow.row,
+});
+
 const pluginDependencies = [
   { pluginName: 'GroupingState' },
   { pluginName: 'TableView' },
@@ -50,6 +56,7 @@ export class TableGroupRow extends React.PureComponent {
   render() {
     const {
       groupCellTemplate,
+      groupRowTemplate,
       groupIndentCellTemplate,
       groupIndentColumnWidth,
     } = this.props;
@@ -105,6 +112,21 @@ export class TableGroupRow extends React.PureComponent {
             )}
           </Template>
         )}
+        <Template
+          name="tableViewRow"
+          predicate={({ tableRow }) => isGroupTableRow(tableRow)}
+        >
+          {params => (
+            <TemplateConnector>
+              {() => (
+                <TemplateRenderer
+                  template={groupRowTemplate}
+                  params={getGroupTableRowTemplateArgs(params)}
+                />
+              )}
+            </TemplateConnector>
+          )}
+        </Template>
       </PluginContainer>
     );
   }
@@ -112,6 +134,7 @@ export class TableGroupRow extends React.PureComponent {
 
 TableGroupRow.propTypes = {
   groupCellTemplate: PropTypes.func.isRequired,
+  groupRowTemplate: PropTypes.func.isRequired,
   groupIndentCellTemplate: PropTypes.func,
   groupIndentColumnWidth: PropTypes.number.isRequired,
 };