+++ packages/dx-react-grid/src/plugins/table-group-row.test.jsx
@@ -7,6 +7,7 @@ import {
   tableRowsWithGrouping,
   isGroupTableCell,
   isGroupIndentTableCell,
+  isGroupTableRow,
 } from '@devexpress/dx-grid-core';
 import { DataTypeProvider } from './data-type-provider';
 import { TableGroupRow } from './table-group-row';
@@ -17,6 +18,7 @@ jest.mock('@devexpress/dx-grid-core', () => ({
   tableRowsWithGrouping: jest.fn(),
   isGroupTableCell: jest.fn(),
   isGroupIndentTableCell: jest.fn(),
+  isGroupTableRow: jest.fn(),
 }));
 
 const defaultDeps = {
@@ -36,6 +38,10 @@ const defaultDeps = {
       tableColumn: { type: 'undefined', id: 1, column: 'column' },
       style: {},
     },
+    tableViewRow: {
+      tableRow: { type: 'undefined', id: 1, row: 'row' },
+      style: {},
+    },
   },
   plugins: ['GroupingState', 'TableView'],
 };
@@ -43,6 +49,7 @@ const defaultDeps = {
 const defaultProps = {
   groupCellTemplate: () => null,
   groupIndentCellTemplate: () => null,
+  groupRowTemplate: () => null,
   groupIndentColumnWidth: 100,
 };
 
@@ -61,6 +68,7 @@ describe('TableGroupRow', () => {
     tableRowsWithGrouping.mockImplementation(() => 'tableRowsWithGrouping');
     isGroupTableCell.mockImplementation(() => false);
     isGroupIndentTableCell.mockImplementation(() => false);
+    isGroupTableRow.mockImplementation(() => false);
   });
   afterEach(() => {
     jest.resetAllMocks();
@@ -195,5 +203,26 @@ describe('TableGroupRow', () => {
       });
     expect(groupCellTemplate.mock.calls[0][0])
       .toHaveProperty('children');
+  });
+
+  it('should render row by using groupRowTemplate', () => {
+    isGroupTableRow.mockImplementation(() => true);
+    const groupRowTemplate = jest.fn(() => null);
+
+    mount(
+      <PluginHost>
+        {pluginDepsToComponents(defaultDeps)}
+        <TableGroupRow
+          {...defaultProps}
+          groupRowTemplate={groupRowTemplate}
+        />
+      </PluginHost>,
+    );
+
+    expect(isGroupTableRow).toBeCalledWith(defaultDeps.template.tableViewRow.tableRow);
+    expect(groupRowTemplate).toBeCalledWith(expect.objectContaining({
+      ...defaultDeps.template.tableViewRow,
+      row: defaultDeps.template.tableViewRow.tableRow.row,
+    }));
   });
 });